+++ src/main/java/com/skillstorm/taxguruplatform/controllers/FormW2Controller.java
@@ -1,21 +1,15 @@
 package com.skillstorm.taxguruplatform.controllers;
 
+import com.skillstorm.taxguruplatform.domain.dtos.FormW2Dto;
+import com.skillstorm.taxguruplatform.exceptions.ForbiddenException;
 import com.skillstorm.taxguruplatform.exceptions.FormW2AlreadyExistsException;
 import com.skillstorm.taxguruplatform.exceptions.FormW2NotFoundException;
+import com.skillstorm.taxguruplatform.services.FormW2Service;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.CrossOrigin;
-import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.PutMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-import com.skillstorm.taxguruplatform.domain.dtos.FormW2Dto;
-import com.skillstorm.taxguruplatform.services.FormW2Service;
+import org.springframework.security.core.Authentication;
+import org.springframework.web.bind.annotation.*;
 
 @RestController
 @RequestMapping("/w2")
@@ -30,18 +24,42 @@ public class FormW2Controller {
     }
 
     @PostMapping
-    public ResponseEntity<FormW2Dto> createFormW2(@RequestBody FormW2Dto formW2Dto) throws FormW2AlreadyExistsException {
+    public ResponseEntity<FormW2Dto> createFormW2(
+            @RequestBody FormW2Dto formW2Dto,
+            @RequestParam String username,
+            Authentication auth)
+            throws FormW2AlreadyExistsException, ForbiddenException {
+        if (!username.equals(auth.getName())) {
+            throw new ForbiddenException("You are not authorized to access this endpoint.");
+        }
+
         return new ResponseEntity<>(formW2Service.create(formW2Dto), HttpStatus.CREATED);
     }
 
     @PutMapping("/{id}")
-    public ResponseEntity<FormW2Dto> fullUpdateFormW2(@PathVariable("id") Long id, @RequestBody FormW2Dto formW2Dto) throws FormW2NotFoundException {
+    public ResponseEntity<FormW2Dto> fullUpdateFormW2(
+            @PathVariable("id") Long id,
+            @RequestParam String username,
+            @RequestBody FormW2Dto formW2Dto,
+            Authentication auth)
+            throws FormW2NotFoundException, ForbiddenException {
+        if (!username.equals(auth.getName())) {
+            throw new ForbiddenException("You are not authorized to access this endpoint.");
+        }
+
         formW2Dto.setId(id);
         return new ResponseEntity<>(formW2Service.fullUpdate(formW2Dto), HttpStatus.OK);
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<Object> deleteFormW2(@PathVariable("id") Long id) throws FormW2NotFoundException {
+    public ResponseEntity<Object> deleteFormW2(
+            @PathVariable("id") Long id,
+            @RequestParam String username,
+            Authentication auth) throws FormW2NotFoundException, ForbiddenException {
+        if (!username.equals(auth.getName())) {
+            throw new ForbiddenException("You are not authorized to access this endpoint.");
+        }
+
         formW2Service.delete(id);
         return new ResponseEntity<>(HttpStatus.NO_CONTENT);
     }
